pyproject.toml
[tool.poetry]
name = "VPC"
version = "0.1.0"
description = ""
authors = ["Zahra-97 <fatimazahra.ouadjane@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.11.0"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py39"]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/setup.py  # exclude a file named foo.py in the root of the project (in addition to the defaults)
'''

[tool.isort]
profile = "black"
skip_gitignore = true
